choose long data type for storing large numbers

+++ hackerrank/src/main/java/com/github/dkoval/algopuzzles/hackerrank/tutorials/ctci/CoinChange.java
@@ -16,18 +16,18 @@ public class CoinChange {
         for (int i = 0; i < m; i++) {
             coins[i] = in.nextInt();
         }
-        int answer = numWaysToMakeChange(n, coins);
+        long answer = numWaysToMakeChange(n, coins);
         System.out.println(answer);
     }
 
-    static int numWaysToMakeChange(int n, int[] coins) {
-        int[][] memo = new int[coins.length][n + 1];
+    static long numWaysToMakeChange(int n, int[] coins) {
+        long[][] memo = new long[coins.length][n + 1];
         for (int i = 0; i < coins.length; i++) {
             memo[i][0] = 1; // only 1 way to reach 0
             for (int amount = 1; amount <= n; amount++) {
                 int coin = coins[i];
-                int numWaysWithoutThisCoin = (i >= 1) ? memo[i - 1][amount] : 0;
-                int numWaysWithThisCoin = (amount >= coin) ? memo[i][amount - coin] : 0;
+                long numWaysWithoutThisCoin = (i >= 1) ? memo[i - 1][amount] : 0;
+                long numWaysWithThisCoin = (amount >= coin) ? memo[i][amount - coin] : 0;
                 memo[i][amount] = numWaysWithoutThisCoin + numWaysWithThisCoin;
             }
         }
